# observability/api-go/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: theta-observability ingest API
  version: "1.0.0"
  description: Multi-tenant observability ingest & control plane.
servers:
  - url: https://api.theta-observability.com
  - url: http://localhost:8080
security:
  - apiKey: []
  - bearer: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Trace:
      type: object
      required: [project_id]
      properties:
        schema_version: { type: string }
        trace_id: { type: string }
        project_id: { type: string }
        name: { type: string }
        run_id: { type: string }
        run_type: { type: string }
        use_case: { type: string }
        user_id: { type: string }
        group: { type: string }
        platform: { type: string }
        model: { type: string }
        status:
          type: string
          enum: [success, error, running]
        tags:
          type: array
          items: { type: string }
        metadata: { type: object }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        latency_ms: { type: integer }
        cost_usd: { type: number }
        token_usage:
          type: object
          properties:
            input: { type: integer }
            output: { type: integer }
            total: { type: integer }
        steps:
          type: array
          items: { $ref: "#/components/schemas/Step" }
        events:
          type: array
          items: { $ref: "#/components/schemas/ObservedEvent" }
    Step:
      type: object
      required: [step_id, type]
      properties:
        step_id: { type: string }
        parent_step_id: { type: string }
        index: { type: integer }
        type:
          type: string
          enum: [llm, tool, retrieval, robotics, human, annotation, custom]
        name: { type: string }
        model: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        latency_ms: { type: integer }
        input_tokens: { type: integer }
        output_tokens: { type: integer }
        messages: { type: array, items: { type: object } }
        tool_calls: { type: array, items: { type: object } }
        events: { type: array, items: { $ref: "#/components/schemas/ObservedEvent" } }
        sensor_frames: { type: array, items: { type: object } }
        metadata: { type: object }
    ObservedEvent:
      type: object
      required: [type]
      properties:
        event_id: { type: string }
        parent_event_id: { type: string }
        step_id: { type: string }
        parent_step_id: { type: string }
        index: { type: integer }
        type: { type: string }
        name: { type: string }
        role: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        latency_ms: { type: integer }
        model: { type: string }
        message: { type: object }
        tool_call: { type: object }
        attachment: { type: object }
        sensor_frame: { type: object }
        value: {}
        metadata: { type: object }
    CorrelationInfo:
      type: object
      properties:
        session_id: { type: string }
        request_id: { type: string }
        parent_trace_id: { type: string }
        root_trace_id: { type: string }
        external_ids: { type: object }
    CanonicalEnvelope:
      type: object
      properties:
        schema_version: { type: string }
        trace_id: { type: string }
        project_id: { type: string }
        name: { type: string }
        source: { type: string }
        kind: { type: string }
        run_id: { type: string }
        run_type: { type: string }
        use_case: { type: string }
        user_id: { type: string }
        session_id: { type: string }
        group: { type: string }
        platform: { type: string }
        model: { type: string }
        status: { type: string }
        tags:
          type: array
          items: { type: string }
        metadata: { type: object }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        latency_ms: { type: integer }
        cost_usd: { type: number }
        correlation: { $ref: "#/components/schemas/CorrelationInfo" }
        events:
          type: array
          items: { $ref: "#/components/schemas/ObservedEvent" }
        attachments:
          type: array
          items: { type: object }
    IngestResponse:
      type: object
      properties:
        trace_id: { type: string }
        ingest_status: { type: string, enum: [accepted] }
        gcs_uri: { type: string }
...
    BulkImportItemResult:
      type: object
      properties:
        index: { type: integer }
        kind: { type: string }
        trace_id: { type: string }
        status: { type: string }
        error: { type: string }
    BulkImportResponse:
      type: object
      properties:
        accepted: { type: integer }
        failed: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/BulkImportItemResult" }
    TraceListItem:
      type: object
      properties:
        trace_id: { type: string }
        project_id: { type: string }
        name: { type: string }
        status: { type: string }
        platform: { type: string }
        model: { type: string }
        user_id: { type: string }
        started_at: { type: string, format: date-time }
        latency_ms: { type: integer }
        total_tokens: { type: integer }
        cost_usd: { type: number }
        step_count: { type: integer }
        has_media: { type: boolean }
        tags: { type: array, items: { type: string } }
    TraceListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/TraceListItem" } }
        next_cursor: { type: string }
...
paths:
...
  /v1/traces:
    post:
      summary: Ingest a trace
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Trace" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestResponse" }
    get:
      summary: List traces
      parameters:
        - in: query
          name: project_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: platform
          schema: { type: string }
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: user_id
          schema: { type: string }
        - in: query
          name: run_id
          schema: { type: string }
        - in: query
          name: run_type
          schema: { type: string }
        - in: query
          name: use_case
          schema: { type: string }
        - in: query
          name: group
          schema: { type: string }
        - in: query
          name: tags
          schema: { type: string }
        - in: query
          name: since
          schema: { type: string, format: date-time }
        - in: query
          name: until
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TraceListResponse" }
...
  /v1/imports/traces:
    post:
      summary: Bulk import traces or canonical envelopes
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                oneOf:
                  - { $ref: "#/components/schemas/Trace" }
                  - { $ref: "#/components/schemas/CanonicalEnvelope" }
          application/x-ndjson:
            schema:
              type: string
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkImportResponse" }
        "207":
          description: Partial success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkImportResponse" }
  /v1/traces/{id}:
    get:
      summary: Get trace meta + raw JSON
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
...